+++ config/routes.rb
@@ -1,4 +1,6 @@
 # frozen_string_literal: true
 
-RailsHealthChecks::Engine.routes.draw do # rubocop:disable Lint/EmptyBlock
+RailsHealthChecks::Engine.routes.draw do
+  get '/',     to: 'health#show', as: :health
+  get '/live', to: 'health#live', as: :health_live
 end